Polish the Event Listener Module
