+++ src/modules/event-listener/event-listener-module-manager.js
@@ -1,16 +1,32 @@
 import BaseModuleManager from '../base-module-manager.js';
 
 class EventListenerModuleManager extends BaseModuleManager {
-    constructor(ctx) {
-        super(ctx);
-        this.ctx = ctx;
-    }
-
-    async initializeAndStartEventListener(eventListenerImplementation) {
+    async initializeBlockchainEventListener(eventListenerImplementation, blockchainConfig) {
         if (this.getImplementation(eventListenerImplementation)) {
             return this.getImplementation(
                 eventListenerImplementation,
-            ).module.initializeAndStartEventListener(this.ctx);
+            ).module.initializeBlockchainEventListener(blockchainConfig);
+        }
+    }
+
+    async getAllPastEvents(
+        eventListenerImplementation,
+        blockchainId,
+        contractName,
+        eventsToFilter,
+        lastCheckedBlock,
+        lastCheckedTimestamp,
+        currentBlock,
+    ) {
+        if (this.getImplementation(eventListenerImplementation)) {
+            return this.getImplementation(eventListenerImplementation).module.getAllPastEvents(
+                blockchainId,
+                contractName,
+                eventsToFilter,
+                lastCheckedBlock,
+                lastCheckedTimestamp,
+                currentBlock,
+            );
         }
     }
